+++ powerauth-data-adapter/src/main/java/io/getlime/security/powerauth/lib/dataadapter/controller/FormDataChangeController.java
@@ -17,18 +17,14 @@ package io.getlime.security.powerauth.lib.dataadapter.controller;
 
 import io.getlime.core.rest.model.base.request.ObjectRequest;
 import io.getlime.core.rest.model.base.response.ObjectResponse;
-import io.getlime.security.powerauth.lib.dataadapter.model.entity.AuthMethodChoiceEntity;
-import io.getlime.security.powerauth.lib.dataadapter.model.entity.BankAccountChoiceEntity;
-import io.getlime.security.powerauth.lib.dataadapter.model.entity.FormDataChangeEntity;
+import io.getlime.security.powerauth.lib.dataadapter.api.DataAdapter;
 import io.getlime.security.powerauth.lib.dataadapter.model.request.FormDataChangeNotificationRequest;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 /**
  * Controller class which handles notifications about changes of operation formData.
@@ -39,6 +35,13 @@ import java.util.logging.Logger;
 @RequestMapping("/api/operation/formData")
 public class FormDataChangeController {
 
+    private DataAdapter dataAdapter;
+
+    @Autowired
+    public FormDataChangeController(DataAdapter dataAdapter) {
+        this.dataAdapter = dataAdapter;
+    }
+
     /**
      * Receive a new operation formData change notification.
      *
@@ -48,28 +51,8 @@ public class FormDataChangeController {
     @RequestMapping(value = "/change", method = RequestMethod.POST)
     public @ResponseBody ObjectResponse formDataChangedNotification(@RequestBody ObjectRequest<FormDataChangeNotificationRequest> request) {
         FormDataChangeNotificationRequest notification = request.getRequestObject();
-        FormDataChangeEntity change = notification.getFormDataChange();
-        switch (change.getType()) {
-            case BANK_ACCOUNT_CHOICE:
-                bankAccountChosen((BankAccountChoiceEntity) change, notification.getUserId(), notification.getOperationId());
-                break;
-            case AUTH_METHOD_CHOICE:
-                authMethodChosen((AuthMethodChoiceEntity) change, notification.getUserId(), notification.getOperationId());
-                break;
-            default:
-                throw new IllegalStateException("Invalid change entity type: " + change.getType());
-        }
+        dataAdapter.formDataChangedNotification(notification);
         return new ObjectResponse();
-    }
-
-    private void bankAccountChosen(BankAccountChoiceEntity bankAccountChoice, String userId, String operationId) {
-        // Handle bank account choice here (e.g. send notification to bank backend).
-        Logger.getLogger(this.getClass().getName()).log(Level.INFO, "Bank account chosen: " + bankAccountChoice.getBankAccountNumber() + ", user: " + userId + ", operationId: " + operationId);
-    }
-
-    private void authMethodChosen(AuthMethodChoiceEntity authMethodChoice, String userId, String operationId) {
-        // Handle authorization method choice here (e.g. send notification to bank backend).
-        Logger.getLogger(this.getClass().getName()).log(Level.INFO, "Authorization method chosen: " + authMethodChoice.getChosenAuthMethod() + ", user: " + userId + ", operationId: " + operationId);
     }
 
 }
