+++ powerauth-data-adapter/src/main/java/io/getlime/security/powerauth/lib/dataadapter/controller/SMSAuthorizationController.java
@@ -17,35 +17,18 @@ package io.getlime.security.powerauth.lib.dataadapter.controller;
 
 import io.getlime.core.rest.model.base.request.ObjectRequest;
 import io.getlime.core.rest.model.base.response.ObjectResponse;
-import io.getlime.security.powerauth.crypto.server.util.DataDigest;
-import io.getlime.security.powerauth.lib.dataadapter.configuration.DataAdapterConfiguration;
+import io.getlime.security.powerauth.lib.dataadapter.api.DataAdapter;
 import io.getlime.security.powerauth.lib.dataadapter.exception.SMSAuthorizationFailedException;
-import io.getlime.security.powerauth.lib.dataadapter.exception.SMSAuthorizationMessageInvalidException;
 import io.getlime.security.powerauth.lib.dataadapter.model.request.CreateSMSAuthorizationRequest;
 import io.getlime.security.powerauth.lib.dataadapter.model.request.VerifySMSAuthorizationRequest;
 import io.getlime.security.powerauth.lib.dataadapter.model.response.CreateSMSAuthorizationResponse;
-import io.getlime.security.powerauth.lib.dataadapter.repository.SMSAuthorizationRepository;
-import io.getlime.security.powerauth.lib.dataadapter.repository.model.entity.SMSAuthorizationEntity;
-import io.getlime.security.powerauth.lib.dataadapter.service.OperationFormDataService;
-import io.getlime.security.powerauth.lib.dataadapter.validation.CreateSMSAuthorizationRequestValidator;
-import io.getlime.security.powerauth.lib.nextstep.model.entity.OperationAmountAttribute;
-import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
-import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Controller;
-import org.springframework.validation.BeanPropertyBindingResult;
-import org.springframework.validation.ValidationUtils;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.math.BigDecimal;
-import java.util.*;
 
 /**
  * Controller class which handles SMS OTP authorization.
@@ -56,21 +39,11 @@ import java.util.*;
 @RequestMapping("/api/auth/sms")
 public class SMSAuthorizationController {
 
-    // the authorization code length - number of digits
-    private static final int AUTHORIZATION_CODE_LENGTH = 8;
-
-    private SMSAuthorizationRepository smsAuthorizationRepository;
-    private DataAdapterConfiguration dataAdapterConfiguration;
-    private OperationFormDataService operationFormDataService;
-    private CreateSMSAuthorizationRequestValidator requestValidator;
+    private DataAdapter dataAdapter;
 
     @Autowired
-    public SMSAuthorizationController(SMSAuthorizationRepository smsAuthorizationRepository, DataAdapterConfiguration dataAdapterConfiguration,
-                                      OperationFormDataService operationFormDataService, CreateSMSAuthorizationRequestValidator requestValidator) {
-        this.smsAuthorizationRepository = smsAuthorizationRepository;
-        this.dataAdapterConfiguration = dataAdapterConfiguration;
-        this.operationFormDataService = operationFormDataService;
-        this.requestValidator = requestValidator;
+    public SMSAuthorizationController(DataAdapter dataAdapter) {
+        this.dataAdapter = dataAdapter;
     }
 
     /**
@@ -80,57 +53,10 @@ public class SMSAuthorizationController {
      * @return Response with message ID.
      */
     @RequestMapping(value = "create", method = RequestMethod.POST)
-    public @ResponseBody ObjectResponse<CreateSMSAuthorizationResponse> create(@RequestBody ObjectRequest<CreateSMSAuthorizationRequest> request) throws MethodArgumentNotValidException {
+    public @ResponseBody ObjectResponse<CreateSMSAuthorizationResponse> create(@RequestBody ObjectRequest<CreateSMSAuthorizationRequest> request) throws MethodArgumentNotValidException, SMSAuthorizationFailedException {
         CreateSMSAuthorizationRequest createSMSAuthorizationRequest = request.getRequestObject();
-
-        // input validation is handled by CreateSMSAuthorizationRequestValidator
-        // validation is invoked manually because of the generified Request object
-        BeanPropertyBindingResult result = new BeanPropertyBindingResult(createSMSAuthorizationRequest, "createSMSAuthorizationRequest");
-        ValidationUtils.invokeValidator(requestValidator, createSMSAuthorizationRequest, result);
-        if (result.hasErrors()) {
-            // getEnclosingMethod() on new object returns a reference to current method
-            MethodParameter methodParam = new MethodParameter(new Object() {
-            }.getClass().getEnclosingMethod(), 0);
-            throw new MethodArgumentNotValidException(methodParam, result);
-        }
-
-        // messageId is generated as random UUID, it can be overriden to provide a real message identification
-        String messageId = UUID.randomUUID().toString();
-
-        // update names of operationData JSON fields if necessary
-        OperationAmountAttribute amountAttribute = operationFormDataService.getAmount(createSMSAuthorizationRequest.getOperationFormData());
-        BigDecimal amount = amountAttribute.getAmount();
-        String currency = amountAttribute.getCurrency();
-        String account = operationFormDataService.getAccount(createSMSAuthorizationRequest.getOperationFormData());
-
-        // update localized SMS message text in resources
-        final DataDigest.Result digestResult = generateAuthorizationCode(amount, currency, account);
-        final String authorizationCode = digestResult.getDigest();
-        final byte[] salt = digestResult.getSalt();
-        String[] messageArgs = {amount.toPlainString(), currency, account, authorizationCode};
-        String messageText = messageSource().getMessage("sms-otp.text", messageArgs, new Locale(createSMSAuthorizationRequest.getLang()));
-
-        SMSAuthorizationEntity smsEntity = new SMSAuthorizationEntity();
-        smsEntity.setMessageId(messageId);
-        smsEntity.setOperationId(createSMSAuthorizationRequest.getOperationId());
-        smsEntity.setUserId(createSMSAuthorizationRequest.getUserId());
-        smsEntity.setOperationName(createSMSAuthorizationRequest.getOperationName());
-        smsEntity.setAuthorizationCode(authorizationCode);
-        smsEntity.setSalt(salt);
-        smsEntity.setMessageText(messageText);
-        smsEntity.setVerifyRequestCount(0);
-        smsEntity.setTimestampCreated(new Date());
-        smsEntity.setTimestampExpires(new DateTime().plusSeconds(dataAdapterConfiguration.getSmsOtpExpirationTime()).toDate());
-        smsEntity.setTimestampVerified(null);
-        smsEntity.setVerified(false);
-
-        // store entity in database
-        smsAuthorizationRepository.save(smsEntity);
-
-        // Add here code to send the SMS OTP message to user identified by userId with messageText.
-
-        CreateSMSAuthorizationResponse createSMSResponse = new CreateSMSAuthorizationResponse(messageId);
-        return new ObjectResponse<>(createSMSResponse);
+        CreateSMSAuthorizationResponse response = dataAdapter.createAuthorizationSMS(createSMSAuthorizationRequest);
+        return new ObjectResponse<>(response);
     }
 
     /**
@@ -140,67 +66,10 @@ public class SMSAuthorizationController {
      * @return Authorization response.
      */
     @RequestMapping(value = "verify", method = RequestMethod.POST)
-    public @ResponseBody ObjectResponse verify(@RequestBody ObjectRequest<VerifySMSAuthorizationRequest> request) throws SMSAuthorizationMessageInvalidException, SMSAuthorizationFailedException {
+    public @ResponseBody ObjectResponse verify(@RequestBody ObjectRequest<VerifySMSAuthorizationRequest> request) throws SMSAuthorizationFailedException {
         VerifySMSAuthorizationRequest verifyRequest = request.getRequestObject();
-        String messageId = verifyRequest.getMessageId();
-        SMSAuthorizationEntity smsEntity = smsAuthorizationRepository.findOne(messageId);
-        if (smsEntity == null) {
-            throw new SMSAuthorizationMessageInvalidException("smsAuthorization.invalidMessage");
-        }
-        // increase number of verification tries and save entity
-        smsEntity.setVerifyRequestCount(smsEntity.getVerifyRequestCount() + 1);
-        smsAuthorizationRepository.save(smsEntity);
-
-        if (smsEntity.getAuthorizationCode() == null || smsEntity.getAuthorizationCode().isEmpty()) {
-            throw new SMSAuthorizationMessageInvalidException("smsAuthorization.invalidCode");
-        }
-        if (smsEntity.isExpired()) {
-            throw new SMSAuthorizationFailedException("smsAuthorization.expired");
-        }
-        if (smsEntity.isVerified()) {
-            throw new SMSAuthorizationFailedException("smsAuthorization.alreadyVerified");
-        }
-        if (smsEntity.getVerifyRequestCount() > dataAdapterConfiguration.getSmsOtpMaxVerifyTriesPerMessage()) {
-            throw new SMSAuthorizationFailedException("smsAuthorization.maxAttemptsExceeded");
-        }
-        String authorizationCodeExpected = smsEntity.getAuthorizationCode();
-        String authorizationCodeActual = verifyRequest.getAuthorizationCode();
-        if (!authorizationCodeActual.equals(authorizationCodeExpected)) {
-            throw new SMSAuthorizationFailedException("smsAuthorization.failed");
-        }
-
-        // SMS OTP authorization succeeded when this line is reached, update entity verification status
-        smsEntity.setVerified(true);
-        smsEntity.setTimestampVerified(new Date());
-        smsAuthorizationRepository.save(smsEntity);
-
-        // no actual data sent - ObjectResponse is empty
+        dataAdapter.verifyAuthorizationSMS(verifyRequest);
         return new ObjectResponse();
     }
 
-    /**
-     * Authorization code generation - to be updated based on application requirements.
-     *
-     * @return Generated authorization code.
-     */
-    private DataDigest.Result generateAuthorizationCode(BigDecimal amount, String currency, String account) {
-        List<String> digestItems = new ArrayList<>();
-        digestItems.add(amount.toPlainString());
-        digestItems.add(currency);
-        digestItems.add(account);
-        return new DataDigest().generateDigest(digestItems);
-    }
-
-    /**
-     * Get MessageSource with i18n data for authorizations SMS messages.
-     *
-     * @return MessageSource.
-     */
-    @Bean
-    private MessageSource messageSource() {
-        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
-        messageSource.setBasename("classpath:/static/resources/messages");
-        messageSource.setDefaultEncoding("UTF-8");
-        return messageSource;
-    }
 }
