+++ powerauth-data-adapter/src/main/java/io/getlime/security/powerauth/lib/dataadapter/impl/validation/AuthenticationRequestValidator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.getlime.security.powerauth.lib.dataadapter.validation;
+package io.getlime.security.powerauth.lib.dataadapter.impl.validation;
 
 import io.getlime.security.powerauth.lib.dataadapter.model.enumeration.AuthenticationType;
 import io.getlime.security.powerauth.lib.dataadapter.model.request.AuthenticationRequest;
